+++ UndeadBits.ValheimMods.PortalStation/PortalStationUser.cs
@@ -1,5 +1,4 @@
-﻿using System;
-using UnityEngine;
+﻿using UnityEngine;
 
 namespace UndeadBits.ValheimMods.PortalStation {
     
@@ -17,12 +16,12 @@ namespace UndeadBits.ValheimMods.PortalStation {
         /// <param name="target">The destination to travel to</param>
         public void Use(PortalStation.Destination from, PortalStation.Destination target) {
             if (!this.view || !this.view.IsValid()) {
-                Jotunn.Logger.LogWarning($"Can't use portal station - player object not valid.");
+                Jotunn.Logger.LogWarning("Can't use portal station - player object not valid.");
                 return;
             }
 
             if (!this.view.IsOwner()) {
-                this.view.InvokeRPC(nameof(RPC_UsePortalStation), from.id, target.id);
+                this.view.InvokeRPC(nameof(RPC_UsePortalStation), from.ID, target.ID);
             } else {
                 UsePortalStation(from, target);
             }
@@ -66,29 +65,29 @@ namespace UndeadBits.ValheimMods.PortalStation {
         /// <param name="target">The destination to travel to</param>
         private void UsePortalStation(PortalStation.Destination station, PortalStation.Destination target) {
             if (!this.player || !this.view) {
-                Jotunn.Logger.LogWarning($"Can't use portal station - player object not valid.");
+                Jotunn.Logger.LogWarning("Can't use portal station - player object not valid.");
                 return;
             }
             
             var playerZdo = this.view.IsValid() ? this.view.GetZDO() : null;
             if (playerZdo == null) {
-                Jotunn.Logger.LogWarning($"Can't use portal station - ZDO not valid.");
+                Jotunn.Logger.LogWarning("Can't use portal station - ZDO not valid.");
                 return;
             }
 
             if (!PortalStationPlugin.Instance.CanTeleportPlayer(this.player)) {
-                Jotunn.Logger.LogWarning($"Can't use portal station - player not teleportable.");
+                Jotunn.Logger.LogWarning("Can't use portal station - player not teleportable.");
                 return;
             }
             
-            var useDistance = Vector3.Distance(playerZdo.GetPosition(), station.position);
+            var useDistance = Vector3.Distance(playerZdo.GetPosition(), station.Position);
             if (useDistance > PortalStation.USE_DISTANCE) {
                 Jotunn.Logger.LogWarning($"Can't use portal station - station {target} out of use range ({useDistance}).");
                 return;
             }
 
-            var position = target.position;
-            var rotation = target.rotation;
+            var position = target.Position;
+            var rotation = target.Rotation;
             var distance = Vector3.Distance(playerZdo.GetPosition(), position);
             var distant = distance >= ZoneSystem.instance.m_zoneSize;